Seperated export to excel functionality
